Revert "Refactor ImmuneCogGraph into modular package"
