--- immune_cog_graph/__init__.py
@@ -1,5 +1,0 @@
-"""Refactored ImmuneCogGraph package."""
-
-from .graph import ImmuneCogGraph
-
-__all__ = ["ImmuneCogGraph"]
