--- immune_cog_graph/cells/__init__.py
@@ -1,11 +1,0 @@
-"""Cell-type specific controllers for the immune cognition graph."""
-
-from .sensor import SensorCellController
-from .processor import ProcessorCellController
-from .emitter import EmitterCellController
-
-__all__ = [
-    "SensorCellController",
-    "ProcessorCellController",
-    "EmitterCellController",
-]
